refactor(main): separate AnkiApiPage
