+++ main/src/entryPoints/desktop/content/index.ts
@@ -1,7 +1,7 @@
 import { TOOLTIP_IFRAME_ID } from "@/consts";
 import { ContentScriptController } from "@/features/content";
 import Utils, { exposeGlobals } from "@/features/utils";
-import { createForegroundDesktopCtx } from "@/platform/desktop";
+import { createContentDesktopCtx } from "@/platform/desktop";
 
 declare global {
   interface Window {
@@ -26,7 +26,7 @@ function maybeInitialize() {
 }
 
 function initialize() {
-  const ctx = createForegroundDesktopCtx();
+  const ctx = createContentDesktopCtx();
 
   exposeGlobals({
     Platform: ctx.platform,
