+++ main/src/entryPoints/desktop/popup/index.ts
@@ -1,11 +1,11 @@
 import type { AppCtx, DesktopCtx } from "@/features/ctx";
 import { Toast } from "@/features/toast";
 import Utils, { exposeGlobals } from "@/features/utils";
-import { createForegroundDesktopCtx } from "@/platform/desktop";
+import { createPageDesktopCtx } from "@/platform/desktop";
 import PopupPage from "./PopupPage.svelte";
 
 async function initialize(): Promise<AppCtx<DesktopCtx>> {
-  const ctx = createForegroundDesktopCtx();
+  const ctx = createPageDesktopCtx();
   const config = await ctx.lazyConfig.get();
   const toast = new Toast(ctx.lazyConfig);
