+++ main/src/platform/desktop/index.ts
@@ -1,4 +1,4 @@
 export { DesktopAnkiApi } from "./anki";
 export { DesktopBackend } from "./backend";
-export { createBackgroundDesktopCtx, createForegroundDesktopCtx } from "./ctx";
+export { createBackgroundDesktopCtx, createContentDesktopCtx, createPageDesktopCtx } from "./ctx";
 export { DesktopPlatform } from "./platform";
